requirements.txt
# the custom library
git+https://github.com/outerbounds/rag-demo

# orchestration
pylint
metaflow

# data processing
gitpython
python-frontmatter
python-slugify
pandas
tldextract

# models
langchain==0.0.262
llama-index==0.8.0
llama-cpp-python==0.1.77
sentence-transformers
openai

# vectorDB
lancedb
pinecone-client[grpc]

# interface
matplotlib
seaborn
notebook
ipywidgets
ipykernel
streamlit
# jupyter_ai # incompatible with latest langchain :/

pyyaml
# NOTE: conflict between pyyaml requirement of python-frontmatter?